+++ src/plugin/gegen_multi.c
@@ -5,7 +5,7 @@
  * Updated: Sat May 20 14:06:40 EDT 2017
  * Purpose: Stata plugin to compute a faster -egen- (multi-threaded version)
  * Note:    See stata.com/plugins for more on Stata plugins
- * Version: 0.3.1
+ * Version: 0.3.2
  *********************************************************************/
 
 #include <omp.h>
@@ -196,8 +196,8 @@ int sf_egen (struct StataInfo *st_info)
         // modify the input data, so the position of each value of the
         // jth group is index[i] for i = start to i < end.
         for (i = start; i < end; i++) {
-            if ( SF_ifobs(st_info->in1 + st_info->index[i]) ) {
-                out = st_info->index[i] + st_info->in1;
+            out = st_info->index[i] + st_info->in1;
+            if ( SF_ifobs(out) ) {
                 if ( (rc = SF_vstore(st_info->start_target_vars, out, output_buffer)) ) return (rc);
             }
         }
@@ -221,14 +221,19 @@ int sf_egen (struct StataInfo *st_info)
  */
 int sf_egen_tag (struct StataInfo *st_info)
 {
+    ST_double z ;
     ST_retcode rc ;
-    int i, j, out;
+    int i, j, k, out;
     size_t start, end, minj;
     clock_t timer = clock();
 
     size_t *indexj = calloc(st_info->J, sizeof *indexj);
     size_t *firstj = calloc(st_info->J, sizeof *firstj);
 
+    // Since we hash the data, the order in C has to be mapped to the
+    // order in Stata via info and index. First figure out the order in
+    // which the groups appear in Stata, and then write by looping over
+    // groups in that order
     for (j = 0; j < st_info->J; j++) {
         start = st_info->info[j];
         end   = st_info->info[j + 1];
@@ -245,20 +250,33 @@ int sf_egen_tag (struct StataInfo *st_info)
         indexj[j] = j;
     }
 
+    // indexj[j] will contain the order in which the jth C group
+    // inappeared Stata
     mf_radix_sort_index (firstj, indexj, st_info->J, RADIX_SHIFT, 0, st_info->verbose);
+    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.1: Tagged groups in memory");
 
+    // We loop in C using indexj and write to Stata based on index
     for (j = 0; j < st_info->J; j++) {
         start = st_info->info[indexj[j]];
         end   = st_info->info[indexj[j] + 1];
         while ( !SF_ifobs(st_info->in1 + st_info->index[start]) & (start < end) ) {
             start++;
         }
-        out = st_info->index[start] + st_info->in1;
         if ( start < end ) {
+            out = st_info->index[start] + st_info->in1;
             if ( (rc = SF_vstore(st_info->start_target_vars, out, 1)) ) return (rc);
         }
     }
-    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5: Tagged groups in Stata");
+
+    // Tag ignores if/in for missing values (all non-tagged are 0)
+    k = st_info->start_target_vars;
+    for (i = 1; i <= SF_nobs(); i++) {
+        if ( (rc = SF_vdata(k, i, &z)) ) return(rc);
+        if ( SF_is_missing(z) ) {
+            if ( (rc = SF_vstore(k, i, 0)) ) return (rc);
+        }
+    }
+    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.2: Copied tag to Stata");
 
     free (indexj);
     free (firstj);
@@ -282,6 +300,10 @@ int sf_egen_group (struct StataInfo *st_info)
     size_t *indexj = calloc(st_info->J, sizeof *indexj);
     size_t *firstj = calloc(st_info->J, sizeof *firstj);
 
+    // Since we hash the data, the order in C has to be mapped to the
+    // order in Stata via info and index. First figure out the order in
+    // which the groups appear in Stata, and then write by looping over
+    // groups in that order
     for (j = 0; j < st_info->J; j++) {
         start = st_info->info[j];
         end   = st_info->info[j + 1];
@@ -298,19 +320,23 @@ int sf_egen_group (struct StataInfo *st_info)
         indexj[j] = j;
     }
 
+    // indexj[j] will contain the order in which the jth C group
+    // inappeared Stata
     mf_radix_sort_index (firstj, indexj, st_info->J, RADIX_SHIFT, 0, st_info->verbose);
+    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.1: Indexed groups in memory");
 
+    // We loop in C using indexj and write to Stata based on index
     for (j = 0; j < st_info->J; j++) {
         start  = st_info->info[indexj[j]];
         end    = st_info->info[indexj[j] + 1];
         for (i = start; i < end; i++) {
-            if ( SF_ifobs(st_info->in1 + st_info->index[i]) ) {
-                out = st_info->index[i] + st_info->in1;
+            out = st_info->index[i] + st_info->in1;
+            if ( SF_ifobs(out) ) {
                 if ( (rc = SF_vstore(st_info->start_target_vars, out, j + 1)) ) return (rc);
             }
         }
     }
-    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5: Indexed groups in Stata");
+    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.2: Copied index to Stata");
 
     free (indexj);
     free (firstj);